Issue #11163: Enforce file size for Indentation
