+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/indentation/InputIndentationCorrectMultipleAnnotationsWithWrappedLines1.java
@@ -1,13 +1,11 @@
 package com.puppycrawl.tools.checkstyle.checks.indentation.indentation; //indent:0 exp:0
 
 
-
-
 @WrappedLinesAnnotation1 //indent:0 exp:0
 @WrappedLinesAnnotation3( //indent:0 exp:0
     "value" //indent:4 exp:4
 ) //indent:0 exp:0
-public class InputIndentationCorrectMultipleAnnotationsWithWrappedLines { //indent:0 exp:0
+public class InputIndentationCorrectMultipleAnnotationsWithWrappedLines1 { //indent:0 exp:0
 
     @WrappedLinesAnnotation1 //indent:4 exp:4
     @WrappedLinesAnnotation3( //indent:4 exp:4
@@ -26,44 +24,22 @@ public class InputIndentationCorrectMultipleAnnotationsWithWrappedLines { //inde
 } //indent:0 exp:0
 
 @WrappedLinesAnnotation1 //indent:0 exp:0
-@WrappedLinesAnnotation3( //indent:0 exp:0
-    "value" //indent:4 exp:4
-) //indent:0 exp:0
-class InputIndentationCorrectMultipleAnnotationsWithWrappedLines2 { //indent:0 exp:0
-
-    @WrappedLinesAnnotation1 //indent:4 exp:4
-    @WrappedLinesAnnotation3( //indent:4 exp:4
-        "value" //indent:8 exp:8
-    ) //indent:4 exp:4
-    public String value; //indent:4 exp:4
-
-    @WrappedLinesAnnotation1 //indent:4 exp:4
-    @WrappedLinesAnnotation3( //indent:4 exp:4
-        "value" //indent:8 exp:8
-    ) //indent:4 exp:4
-    public String method() { //indent:4 exp:4
-        return "value"; //indent:8 exp:8
-    }; //indent:4 exp:4
-
-} //indent:0 exp:0
-
-@WrappedLinesAnnotation1 //indent:0 exp:0
-@WrappedLinesAnnotation3( //indent:0 exp:0
-    "value" //indent:4 exp:4
+@WrappedLinesAnnotation4( //indent:0 exp:0
+    { "value" } //indent:4 exp:4
 ) //indent:0 exp:0
 @WrappedLinesAnnotation2 //indent:0 exp:0
-class InputIndentationCorrectMultipleAnnotationsWithWrappedLines3 { //indent:0 exp:0
+class InputIndentationCorrectMultipleAnnotationsWithWrappedLines6 { //indent:0 exp:0
 
     @WrappedLinesAnnotation1 //indent:4 exp:4
-    @WrappedLinesAnnotation3( //indent:4 exp:4
-        "value" //indent:8 exp:8
+    @WrappedLinesAnnotation4( //indent:4 exp:4
+        { "value" } //indent:8 exp:8
     ) //indent:4 exp:4
     @WrappedLinesAnnotation2 //indent:4 exp:4
     public String value; //indent:4 exp:4
 
     @WrappedLinesAnnotation1 //indent:4 exp:4
-    @WrappedLinesAnnotation3( //indent:4 exp:4
-        "value" //indent:8 exp:8
+    @WrappedLinesAnnotation4( //indent:4 exp:4
+        { "value" } //indent:8 exp:8
     ) //indent:4 exp:4
     @WrappedLinesAnnotation2 //indent:4 exp:4
     public String method() { //indent:4 exp:4
@@ -73,45 +49,65 @@ class InputIndentationCorrectMultipleAnnotationsWithWrappedLines3 { //indent:0 e
 } //indent:0 exp:0
 
 @WrappedLinesAnnotation1 //indent:0 exp:0
-@WrappedLinesAnnotation3( //indent:0 exp:0
+@WrappedLinesAnnotation4({ //indent:0 exp:0
     "value" //indent:4 exp:4
-) @WrappedLinesAnnotation2 //indent:0 exp:0
-class InputIndentationCorrectMultipleAnnotationsWithWrappedLines4 { //indent:0 exp:0
+}) //indent:0 exp:0
+@WrappedLinesAnnotation2 //indent:0 exp:0
+class InputIndentationCorrectMultipleAnnotationsWithWrappedLines7 { //indent:0 exp:0
 
     @WrappedLinesAnnotation1 //indent:4 exp:4
-    @WrappedLinesAnnotation3( //indent:4 exp:4
+    @WrappedLinesAnnotation4({ //indent:4 exp:4
         "value" //indent:8 exp:8
-    ) @WrappedLinesAnnotation2 //indent:4 exp:4
+    }) //indent:4 exp:4
+    @WrappedLinesAnnotation2 //indent:4 exp:4
     public String value; //indent:4 exp:4
 
     @WrappedLinesAnnotation1 //indent:4 exp:4
-    @WrappedLinesAnnotation3( //indent:4 exp:4
+    @WrappedLinesAnnotation4({ //indent:4 exp:4
         "value" //indent:8 exp:8
-    ) @WrappedLinesAnnotation2 //indent:4 exp:4
+    }) //indent:4 exp:4
+    @WrappedLinesAnnotation2 //indent:4 exp:4
     public String method() { //indent:4 exp:4
         return "value"; //indent:8 exp:8
     }; //indent:4 exp:4
 
 } //indent:0 exp:0
 
-@WrappedLinesAnnotation3( //indent:0 exp:0
-    "value" //indent:4 exp:4
+@WrappedLinesAnnotation1 //indent:0 exp:0
+@WrappedLinesAnnotationOuter( //indent:0 exp:0
+    @WrappedLinesAnnotationInner //indent:4 exp:4
 ) //indent:0 exp:0
 @WrappedLinesAnnotation2 //indent:0 exp:0
-class InputIndentationCorrectMultipleAnnotationsWithWrappedLines5 { //indent:0 exp:0
+//indent:0 exp:0
+class InputIndentationCorrectMultipleAnnotationsWithWrappedLines8 { //indent:0 exp:0
 
-    @WrappedLinesAnnotation3( //indent:4 exp:4
-        "value" //indent:8 exp:8
+    @WrappedLinesAnnotation1 //indent:4 exp:4
+    @WrappedLinesAnnotationOuter( //indent:4 exp:4
+        @WrappedLinesAnnotationInner //indent:8 exp:8
     ) //indent:4 exp:4
     @WrappedLinesAnnotation2 //indent:4 exp:4
     public String value; //indent:4 exp:4
 
-    @WrappedLinesAnnotation3( //indent:4 exp:4
-        "value" //indent:8 exp:8
+    @WrappedLinesAnnotation1 //indent:4 exp:4
+    @WrappedLinesAnnotationOuter( //indent:4 exp:4
+        @WrappedLinesAnnotationInner //indent:8 exp:8
     ) //indent:4 exp:4
     @WrappedLinesAnnotation2 //indent:4 exp:4
     public String method() { //indent:4 exp:4
         return "value"; //indent:8 exp:8
     }; //indent:4 exp:4
 
+} //indent:0 exp:0
+
+@interface WrappedLinesAnnotation1 {} //indent:0 exp:0
+@interface WrappedLinesAnnotation2 {} //indent:0 exp:0
+@interface WrappedLinesAnnotation3 { //indent:0 exp:0
+    String value(); //indent:4 exp:4
+} //indent:0 exp:0
+@interface WrappedLinesAnnotation4 { //indent:0 exp:0
+    String[] value(); //indent:4 exp:4
+} //indent:0 exp:0
+@interface WrappedLinesAnnotationInner {} //indent:0 exp:0
+@interface WrappedLinesAnnotationOuter { //indent:0 exp:0
+    WrappedLinesAnnotationInner[] value(); //indent:4 exp:4
 } //indent:0 exp:0
