+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/indentation/InputIndentationInvalidArrayInitIndent.java
@@ -76,61 +76,4 @@ public class InputIndentationInvalidArrayInitIndent { //indent:0 exp:0
           new int[] { 1, 2, 3},
             new int[] { 1, 2, 3},
         };
-
-    private void func1(int[] arg) {
-
-    }
-
-
-    /** Creates a new instance of InputIndentationValidArrayInitIndent */
-    public InputIndentationInvalidArrayInitIndent() {
-
-        func1(new int[] {
-        1, 2, 3
-        });
-    }
-
-    private static char[] sHexChars = {
-        '0', '1', '2', '3', '4', '5', '6', '7',
-        '8', '9', 'A', 'B', 'C', 'D', 'E', 'F' };
-
-    private void myFunc3()
-    {
-        char[] sHexChars2 = {
-          '0', '1', '2', '3', '4', '5', '6', '7',
-              '8', '9', 'A', 'B', 'C', 'D', 'E', 'F' };
-
-        char[] sHexChars3 = {
-          '0', '1', '2', '3', '4', '5', '6', '7',
-              '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'
-      };
-
-        char[] sHexChars4 =
-      {
-              '0', '1', '2', '3', '4', '5', '6', '7',
-          '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'
-      };
-
-    }
-    interface SomeInterface1 {
-        @interface SomeAnnotation1 {
-            String[] values();
-            String[] description() default "";
-            String[] description2() default { "hello",
-            "checkstyle"};
-        }
-        @SomeAnnotation1(values =
-            {
-              "hello",
-            "checkstyle"
-            },
-            description = { "hello",
-              "checkstyle"
-        },
-            description2 = {
-            "hello", "chekstyle"
-            }
-        )
-      void worked();
-    }
-}
+} //indent:0 exp:0
