+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/indentation/InputIndentationInvalidBlockIndent.java
@@ -88,66 +88,4 @@ public class InputIndentationInvalidBlockIndent { //indent:0 exp:0
 
     } //indent:4 exp:4
 
-
-//  static init at beginning of line is broken for now //indent:0 exp:0
-
-
-  static { int var = 4; } //indent:2 exp:4 warn
-      static { int var = 4; } //indent:6 exp:4 warn
-
-
-    static { //indent:4 exp:4
-       int var = 4;  //indent:7 exp:8 warn
-    } //indent:4 exp:4
-
-      static { //indent:6 exp:4 warn
-        int var = 4;  //indent:8 exp:8
-  } //indent:2 exp:4 warn
-
-  static { //indent:2 exp:4 warn
-        int var = 4;  //indent:8 exp:8
-      } //indent:6 exp:4 warn
-
-  static  //indent:2 exp:4 warn
-    { //indent:4 exp:4
-      int var = 4;  //indent:6 exp:8 warn
-    } //indent:4 exp:4
-    static  //indent:4 exp:4
-  { //indent:2 exp:4 warn
-      int var = 4;  //indent:6 exp:8 warn
-      } //indent:6 exp:4 warn
-
-
-    static  //indent:4 exp:4
-    { //indent:4 exp:4
-      int var = 4;  //indent:6 exp:8 warn
-    } //indent:4 exp:4
-
-    static  //indent:4 exp:4
-    { //indent:4 exp:4
-    int var = 4;  //indent:4 exp:8 warn
-  } //indent:2 exp:4 warn
-
-    static  //indent:4 exp:4
-    { //indent:4 exp:4
-        int var = 4;  //indent:8 exp:8
-      } //indent:6 exp:4 warn
-
-
-  { int var = 4; } //indent:2 exp:4 warn
-      { int var = 4; } //indent:6 exp:4 warn
-
-
-  { //indent:2 exp:4 warn
-        int var = 4;  //indent:8 exp:8
-      } //indent:6 exp:4 warn
-
-      { //indent:6 exp:4 warn
-        int var = 4;  //indent:8 exp:8
-  } //indent:2 exp:4 warn
-
-    { //indent:4 exp:4
-      int var = 4;  //indent:6 exp:8 warn
-    } //indent:4 exp:4
-
 } //indent:0 exp:0
