+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/indentation/InputIndentationInvalidIfIndent.java
@@ -106,162 +106,4 @@ public class InputIndentationInvalidIfIndent { //indent:0 exp:0
            } //indent:11 exp:8 warn
     } //indent:4 exp:4
 
-    /////  same as above, with statements //indent:4 exp:4
-    public void populatedIfTest() //indent:4 exp:4
-    { //indent:4 exp:4
-        boolean test = false; //indent:8 exp:8
-        // no braces if //indent:8 exp:8
-        if (test) //indent:8 exp:8
-              System.getProperty("blah"); //indent:14 exp:>=12
-
-        // no braces if/else //indent:8 exp:8
-        if (test) //indent:8 exp:8
-            System.getProperty("blah"); //indent:12 exp:>=12
-        else //indent:8 exp:8
-            System.getProperty("blah"); //indent:12 exp:>=12
-
-
-        // lcurly on same line, and stmt //indent:8 exp:8
-        if (test) { //indent:8 exp:8
-              System.getProperty("blah"); //indent:14 exp:12 warn
-        } //indent:8 exp:8
-
-        // lcurly on next line and stmt //indent:8 exp:8
-        if (test)  //indent:8 exp:8
-          { //indent:10 exp:8 warn
-          System.getProperty("blah"); //indent:10 exp:12 warn
-        } //indent:8 exp:8
-        // lcurly for if and else on same line //indent:8 exp:8
-        if (test) { //indent:8 exp:8
-
-              System. //indent:14 exp:12 warn
-          getProperty("blah"); //indent:10 exp:12 warn
-        } else { //indent:8 exp:8
-          System. //indent:10 exp:12 warn
-        getProperty("blah"); //indent:8 exp:12 warn
-        } //indent:8 exp:8
-
-        // lcurly for if and else on same line //indent:8 exp:8
-        if (test)  //indent:8 exp:8
-        { //indent:8 exp:8
-            System.getProperty("blah"); //indent:12 exp:12
-                System.getProperty("blah"); //indent:16 exp:12 warn
-         }  //indent:9 exp:8 warn
-        else  //indent:8 exp:8
-        { //indent:8 exp:8
-                System.getProperty("blah"); //indent:16 exp:12 warn
-            System.getProperty("blah"); //indent:12 exp:12
-        } //indent:8 exp:8
-
-        // lcurly for if and else on same line -- mixed braces //indent:8 exp:8
-        if (test) { //indent:8 exp:8
-System.getProperty("blah"); //indent:0 exp:12 warn
-        }  //indent:8 exp:8
-        else  //indent:8 exp:8
-        { //indent:8 exp:8
-                                        System.getProperty("blah"); //indent:40 exp:12 warn
-        } //indent:8 exp:8
-
-
-        // lcurly for if and else on same line -- mixed braces //indent:8 exp:8
-        if (test)  //indent:8 exp:8
-        { //indent:8 exp:8
-              System.getProperty("blah"); //indent:14 exp:12 warn
-        } else  //indent:8 exp:8
-        { //indent:8 exp:8
-              System.getProperty("blah"); //indent:14 exp:12 warn
-        } //indent:8 exp:8
-
-        // lcurly for if and else on same line -- mixed braces //indent:8 exp:8
-        if (test)  //indent:8 exp:8
-        { //indent:8 exp:8
-          System.getProperty("blah"); //indent:10 exp:12 warn
-        } else { //indent:8 exp:8
-          System.getProperty("blah"); //indent:10 exp:12 warn
-        } //indent:8 exp:8
-
-        // lcurly for if and else on same line -- mixed braces, unnested //indent:8 exp:8
-          if (test) { //indent:10 exp:8 warn
-              System.getProperty("blah"); //indent:14 exp:12 warn
-          }  //indent:10 exp:8 warn
-          else { //indent:10 exp:8 warn
-              System.getProperty("blah"); //indent:14 exp:12 warn
-          } //indent:10 exp:8 warn
-
-        if (test //indent:8 exp:8
-         && 7 < 8 && 8 < 9 //indent:9 exp:12 warn
-           && 10 < 11) { //indent:11 exp:12 warn
-        } //indent:8 exp:8
-
-        if (test) //indent:8 exp:8
-          return; //indent:10 exp:12 warn
-
-        if (test) { //indent:8 exp:8
-       } else if (7 < 8) { //indent:7 exp:8 warn
-        } else if (8 < 9) { //indent:8 exp:8
-        } //indent:8 exp:8
-
-        if (test) { //indent:8 exp:8
-            System.getProperty("blah");  //indent:12 exp:>=12
-        } else if (7 < 8) { //indent:8 exp:8
-          System.getProperty("blah");  //indent:10 exp:>=12 warn
-        } else if (8 < 9) { //indent:8 exp:8
-          System.getProperty("blah");  //indent:10 exp:>=12 warn
-        } //indent:8 exp:8
-
-
-        if (test) //indent:8 exp:8
-            System.getProperty("blah");  //indent:12 exp:12
-        else if (7 < 8) //indent:8 exp:8
-          System.getProperty("blah");  //indent:10 exp:12 warn
-        else if (8 < 9) //indent:8 exp:8
-            System.getProperty("blah");  //indent:12 exp:12
-
-
-        //                               //indent:8 exp:8
-        if (test) { //indent:8 exp:8
-            System.getProperty("blah");  //indent:12 exp:12
-        } else  //indent:8 exp:8
-          if (7 < 8) { //indent:10 exp:12 warn
-                System.getProperty("blah");  //indent:16 exp:16
-            } else  //indent:12 exp:12
-                if (8 < 9) { //indent:16 exp:16
-                  System.getProperty("blah");  //indent:18 exp:20 warn
-                } //indent:16 exp:16
-
-        if (test) { //indent:8 exp:8
-            System.getProperty("blah"); } //indent:12 exp:12
-    } //indent:4 exp:4
-
-    public void parenIfTest() { //indent:4 exp:4
-        boolean test = true; //indent:8 exp:8
-
-        if (test //indent:8 exp:8
-          ) { //indent:10 exp:8 warn
-            System.getProperty("blah");  //indent:12 exp:12
-        } //indent:8 exp:8
-
-        if (test //indent:8 exp:8
-      )  //indent:6 exp:8 warn
-        { //indent:8 exp:8
-            System.getProperty("blah");  //indent:12 exp:12
-        } //indent:8 exp:8
-
-        if  //indent:8 exp:8
-      ( //indent:6 exp:8 warn
-            test //indent:12 exp:12
-      ) { //indent:6 exp:8 warn
-            System.getProperty("blah");  //indent:12 exp:12
-        } //indent:8 exp:8
-if (test  //indent:0 exp:8 warn
-|| test) {  //indent:0 exp:12 warn
-System.getProperty("blah");  //indent:0 exp:12 warn
-}  //indent:0 exp:8 warn
-if (test) //indent:0 exp:8 warn
-System.getProperty("blah"); //indent:0 exp:12 warn
-else //indent:0 exp:8 warn
-System.getProperty("blah"); //indent:0 exp:12 warn
-
-    } //indent:4 exp:4
-
 } //indent:0 exp:0
