+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/indentation/InputIndentationLabels1.java
@@ -14,56 +14,9 @@ package com.puppycrawl.tools.checkstyle.checks.indentation.indentation; //indent
  *                                                                            //indent:1 exp:1
  *                                                                            //indent:1 exp:1
  */                                                                           //indent:1 exp:1
-class InputIndentationLabels { //indent:0 exp:0
-  void foo() { //indent:2 exp:2
-//    OUT: //indent:0 exp:0
-    while (true) { //indent:4 exp:4
-    } //indent:4 exp:4
-  } //indent:2 exp:2
+public class InputIndentationLabels1 { //indent:0 exp:0
 
-  void foo2() { //indent:2 exp:2
-    positions: while (true) { //indent:4 exp:4
-    } //indent:4 exp:4
-  } //indent:2 exp:2
-
-  void foo3() { //indent:2 exp:2
-    OUT1: //indent:4 exp:4
-    for (;;) { //indent:4 exp:4
-      if (true){ //indent:6 exp:6
-        break OUT1; //indent:8 exp:8
-      } //indent:6 exp:6
-    } //indent:4 exp:4
-  } //indent:2 exp:2
-
-  void foo4() { //indent:2 exp:2
-    OUT1: for (;;) { //indent:4 exp:4
-      if (true){ //indent:6 exp:6
-        break OUT1; //indent:8 exp:8
-      } //indent:6 exp:6
-    } //indent:4 exp:4
-  } //indent:2 exp:2
-
-  void fooo() { //indent:2 exp:2
-    IN: if (true) { //indent:4 exp:4
-    } //indent:4 exp:4
-  } //indent:2 exp:2
-
-  void fooo1() { //indent:2 exp:2
-    IN: //indent:4 exp:4
-    if (true) { //indent:4 exp:4
-    } //indent:4 exp:4
-  } //indent:2 exp:2
-
-  void foooo() { //indent:2 exp:2
-    IN: do {} while (true); //indent:4 exp:4
-  } //indent:2 exp:2
-
-  void foooo1() { //indent:2 exp:2
-    IN: //indent:4 exp:4
-    do {} while (true); //indent:4 exp:4
-  } //indent:2 exp:2
-
-  class Inner { //indent:2 exp:2
+  InputIndentationLabels anon = new InputIndentationLabels() { //indent:2 exp:2
     void foo() { //indent:4 exp:4
       OUT: while (true) { //indent:6 exp:6
       } //indent:6 exp:6
@@ -111,5 +64,5 @@ class InputIndentationLabels { //indent:0 exp:0
       IN: //indent:6 exp:6
       do {} while (true); //indent:6 exp:6
     } //indent:4 exp:4
-  } //indent:2 exp:2
+  }; //indent:2 exp:2
 } //indent:0 exp:0
