+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/indentation/InputIndentationValidIfIndent1.java
@@ -12,99 +12,91 @@ package com.puppycrawl.tools.checkstyle.checks.indentation.indentation; //indent
  * tabWidth = 4                                                               //indent:1 exp:1
  * throwsIndent = 4                                                           //indent:1 exp:1
  *                                                                            //indent:1 exp:1
- * @author  jrichard                                                         //indent:1 exp:1
+ * @author  jrichard                                                          //indent:1 exp:1
  */                                                                           //indent:1 exp:1
-public class InputIndentationValidIfIndent { //indent:0 exp:0
+public class InputIndentationValidIfIndent1 { //indent:0 exp:0
 
-    public void emptyIfTest() //indent:4 exp:4
+    /////  same as above, with statements //indent:4 exp:4
+    public void populatedIfTest() //indent:4 exp:4
     { //indent:4 exp:4
-        boolean test = true; //indent:8 exp:8
+        boolean test = false; //indent:8 exp:8
+        // no braces if //indent:8 exp:8
+        if (test) //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
 
-        // lcurly on same line  //indent:8 exp:8
+        // no braces if/else //indent:8 exp:8
+        if (test) //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
+        else //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
+
+
+        // lcurly on same line, and stmt //indent:8 exp:8
         if (test) { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
 
-        // lcurly on next line //indent:8 exp:8
+        // lcurly on next line and stmt //indent:8 exp:8
         if (test) //indent:8 exp:8
         { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
-
         // lcurly for if and else on same line //indent:8 exp:8
         if (test) { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } else { //indent:8 exp:8
+            System. //indent:12 exp:12
+                getProperty("blah"); //indent:16 exp:>=16
         } //indent:8 exp:8
 
         // lcurly for if and else on same line //indent:8 exp:8
         if (test) //indent:8 exp:8
         { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
         else //indent:8 exp:8
         { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
 
         // lcurly for if and else on same line -- mixed braces //indent:8 exp:8
         if (test) { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
         else //indent:8 exp:8
         { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
 
 
         // lcurly for if and else on same line -- mixed braces //indent:8 exp:8
         if (test) //indent:8 exp:8
         { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } else //indent:8 exp:8
         { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
 
         // lcurly for if and else on same line -- mixed braces //indent:8 exp:8
         if (test) //indent:8 exp:8
         { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } else { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
 
         // lcurly for if and else on same line -- mixed braces, unnested //indent:8 exp:8
         if (test) { //indent:8 exp:8
+            System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
         else { //indent:8 exp:8
-        } //indent:8 exp:8
-
-        if (foo() //indent:8 exp:8
-            || test //indent:12 exp:12
-            || test) //indent:12 exp:12
-        { //indent:8 exp:8
-        } //indent:8 exp:8
-
-    } //indent:4 exp:4
-
-    public void parenIfTest() { //indent:4 exp:4
-        boolean test = true; //indent:8 exp:8
-
-        if (test //indent:8 exp:8
-        ) { //indent:8 exp:8
             System.getProperty("blah"); //indent:12 exp:12
         } //indent:8 exp:8
 
-        if (test //indent:8 exp:8
-        ) //indent:8 exp:8
-        { //indent:8 exp:8
-            System.getProperty("blah"); //indent:12 exp:12
-        } //indent:8 exp:8
+        if (test) System.getProperty("blah"); //indent:8 exp:8
 
-        if //indent:8 exp:8
-        ( //indent:8 exp:12 warn
-            test //indent:12 exp:12
-        ) { //indent:8 exp:8
-            System.getProperty("blah"); //indent:12 exp:12
-        } //indent:8 exp:8
-
-        if (test //indent:8 exp:8
-            ) //indent:12 exp:12
-        { //indent:8 exp:8
-        } //indent:8 exp:8
     } //indent:4 exp:4
 
-    boolean foo() { //indent:4 exp:4
-        return true; //indent:8 exp:8
-    } //indent:4 exp:4
 } //indent:0 exp:0
